Moved workspace-item
