+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/presentation/workspaces-items/workspaces-items.tsx
@@ -15,7 +15,7 @@
 import * as React from 'react'
 import styled from 'styled-components'
 import { CustomElement } from '../../styles/mixins'
-import WorkspaceItemContainer from '../../container/workspace-item-container'
+import WorkspaceItemContainer from '../../workspace-item'
 import { Button, buttonTypeEnum } from '../button'
 import { hot } from 'react-hot-loader'
